Correcoes de ajustes
Removido o  UpdateRTCbyTS, (sera implementado melhor na proxima att).
Calculo da semana foi para global.
