+++ src/RtcDateTime.cpp
@@ -21,21 +21,6 @@ uint8_t StringToUint8(const char *pString)
   return value;
 }
 
-uint8_t dow(uint16_t y, uint8_t m, uint8_t d)
-{
-  uint8_t dow;
-  uint8_t dowArray[] PROGMEM = {0, 3, 2, 5, 0, 3, 5, 1, 4, 6, 2, 4};
-
-  y -= m < 3;
-  dow = ((y + y / 4 - y / 100 + y / 400 + pgm_read_byte(dowArray + (m - 1)) + d) % 7);
-
-  if (dow == 0)
-  {
-    return 7;
-  }
-  return dow;
-}
-
 DateTime::DateTime(const char *date, const char *time)
 {
   // sample input: date = "Dec 06 2009", time = "12:34:56"
